% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexCalculator.R
\name{IndexCalculator}
\alias{IndexCalculator}
\title{IndexCalculator: Computes acoustic indices from a wave object}
\description{
IndexCalculator: Computes acoustic indices from a wave object

IndexCalculator: Computes acoustic indices from a wave object
}
\details{
This class encapsulates the logic for computing multiple ecoacoustic indices
from a stereo WAV file, using the \code{soundecology} and \code{seewave} packages.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{Path of the WAV file}

\item{\code{wav}}{A \code{tuneR::Wave} object}

\item{\code{logger}}{Optional logger object for error reporting}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{filename}}{Path of the WAV file}

\item{\code{wav}}{A \code{tuneR::Wave} object}

\item{\code{logger}}{Optional logger object for error reporting}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IndexCalculator-new}{\code{IndexCalculator$new()}}
\item \href{#method-IndexCalculator-measure_time}{\code{IndexCalculator$measure_time()}}
\item \href{#method-IndexCalculator-compute_indices}{\code{IndexCalculator$compute_indices()}}
\item \href{#method-IndexCalculator-duration}{\code{IndexCalculator$duration()}}
\item \href{#method-IndexCalculator-aci}{\code{IndexCalculator$aci()}}
\item \href{#method-IndexCalculator-ndsi}{\code{IndexCalculator$ndsi()}}
\item \href{#method-IndexCalculator-bio}{\code{IndexCalculator$bio()}}
\item \href{#method-IndexCalculator-adi}{\code{IndexCalculator$adi()}}
\item \href{#method-IndexCalculator-aei}{\code{IndexCalculator$aei()}}
\item \href{#method-IndexCalculator-entropy}{\code{IndexCalculator$entropy()}}
\item \href{#method-IndexCalculator-temp_entropy}{\code{IndexCalculator$temp_entropy()}}
\item \href{#method-IndexCalculator-spec_entropy}{\code{IndexCalculator$spec_entropy()}}
\item \href{#method-IndexCalculator-mae}{\code{IndexCalculator$mae()}}
\item \href{#method-IndexCalculator-np}{\code{IndexCalculator$np()}}
\item \href{#method-IndexCalculator-clone}{\code{IndexCalculator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-new"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$new(filename, wav, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{File path of the original audio}

\item{\code{wav}}{The wave object (tuneR::Wave)}

\item{\code{logger}}{Optional logger for error reporting
Measure execution time of a function and attach it to the result}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-measure_time"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-measure_time}{}}}
\subsection{Method \code{measure_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$measure_time(func, index_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func}}{The function to execute}

\item{\code{index_name}}{Name of the index (used for naming the timing field)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list with results and processing time
Compute the selected acoustic indices
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-compute_indices"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-compute_indices}{}}}
\subsection{Method \code{compute_indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$compute_indices(indices = private$all_indices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{Vector of index names to compute (default: all)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble with the results
Get the duration of the WAV file
Compute Acoustic Complexity Index (ACI)
Compute Normalized Difference Soundscape Index (NDSI)
Compute Bioacoustic Index (BIO)
Compute Acoustic Diversity Index (ADI)
Compute Acoustic Evenness Index (AEI)
Compute Entropy Index (Shannon Entropy)
Compute Temporal Entropy
Compute Spectral Entropy
Compute Amplitude Envelope Mean (MAE)
Compute Number of Peaks (NP) in spectrum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-duration"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-duration}{}}}
\subsection{Method \code{duration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$duration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-aci"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-aci}{}}}
\subsection{Method \code{aci()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$aci()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-ndsi"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-ndsi}{}}}
\subsection{Method \code{ndsi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$ndsi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-bio"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-bio}{}}}
\subsection{Method \code{bio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$bio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-adi"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-adi}{}}}
\subsection{Method \code{adi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$adi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-aei"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-aei}{}}}
\subsection{Method \code{aei()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$aei()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-entropy"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-entropy}{}}}
\subsection{Method \code{entropy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$entropy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-temp_entropy"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-temp_entropy}{}}}
\subsection{Method \code{temp_entropy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$temp_entropy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-spec_entropy"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-spec_entropy}{}}}
\subsection{Method \code{spec_entropy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$spec_entropy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-mae"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-mae}{}}}
\subsection{Method \code{mae()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$mae()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-np"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-np}{}}}
\subsection{Method \code{np()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$np()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IndexCalculator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IndexCalculator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexCalculator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
